refactor: rename inner.rs to algo.rs
